requirements.txt
deepface==0.0.75
numpy==1.21.6
opencv_python==4.3.0.36
Pillow==9.1.1
streamlit==1.10.0
